refact Permissions and GetGroupPermissions service
